Updates output filename convention to match svcomp
svcomp wants "true" in the file name if it has no bug, false if it does.

I originally wrote the tool chain backwards.  This commit renames the
variables that store the buginess value (from "Bug" to "NoBug"), so it
is more clear that value being used in the output filename should
indicate true if file has no bug, and false if it does.