FindEntryPointsVisitor - Fix AST Matching
After correcting the function pointer type of _whoop_init to match
the signature of the actual driver init functions, the ast structure
of the _whoop_init declaration/definition changed because now it doesn't
have to do a cast to change the signature of the function pointer to
match the function being shoved into it.

As an example,

|-VarDecl 0x47909c0 prev 0x458ce30 <line:9296:1, col:30> col:9 _whoop_init 'void (*)(void)' cinit
| `-ImplicitCastExpr 0x4790a58 <col:30> 'void (*)(void)' <BitCast>
|   `-ImplicitCastExpr 0x4790a40 <col:30> 'int (*)(void)' <FunctionToPointerDecay>
|     `-DeclRefExpr 0x4790a18 <col:30> 'int (void)' Function 0x478fe50 'nvram_init' 'int (void)'

becomes

|-VarDecl 0x51af5b0 prev 0x4fabe30 <line:9296:1, col:29> col:8 _whoop_init 'int (*)(void)' cinit
| `-ImplicitCastExpr 0x51af630 <col:29> 'int (*)(void)' <FunctionToPointerDecay>
|   `-DeclRefExpr 0x51af608 <col:29> 'int (void)' Function 0x51aea40 'nvram_init' 'int (void)'

once the function pointer signature matches that of the function being
pointed to.